+++ dimabe_jwt_token/api_env.py
@@ -1,0 +1,3 @@
+secret = '39029eb1-d41a-4b5d-b2c6-40aa94ff5acc'
+db = 'dimabe-odoo-la-invernada-dev-801206'
+algorithm = 'HS256'
