--- dimabe_jwt_token/models/secret_code_config.py
@@ -1,13 +1,0 @@
-from odoo import fields, models, api
-
-
-class SecretCodeConfig(models.TransientModel):
-    _name = 'secret.code.config'
-    _inherit = 'res.config.settings'
-
-    api_secret_code = fields.Text(
-        'Código Secreto'
-    )
-
-
-
